Deleted unnecessary files that were made in the init
including the NotificationsFragment.java and xml
and DashboardFragment.java and xml plus any references made to that in the "ui" folder and the mobile_navigation.xml graph.

I also cleaned the project and rebuilt.
then added fragment_home.xml and HomeFragment.java
HomeFragment.java will be used as model for the home
and saved screens. We will do this by making HomeFragment.java hold the data and simply switch
between fragment_home.xml and fragment_saved.xml

## Added new files to support the wireframe

MainActivity is the main controller that manages everything on the screen. It shows different sections like "Home" and "Saved" by using smaller pieces called fragments.

HomeFragment and SavedFragment are listed in a navigation graph file called mobile_navigation.xml. This navigation graph is what the app uses to keep track of which section (or fragment) should be shown at any time.

In activity_main.xml, there is something called a NavHostFragment which works like a container. It displays whichever fragment is needed at the moment, based on what the user selects. The BottomNavigationView is the menu at the bottom that lets users switch between "Home" and "Saved". When the user clicks a button in this menu, the NavHostFragment changes to show the right section.

Overall, MainActivity manages everything and decides which section to display, while HomeFragment and SavedFragment are the individual screens that users interact with. The navigation graph (mobile_navigation.xml) just tells the app which fragments are available and how to switch between them.